Divide services in tags
