+++ kraftwerk-api/src/main/resources/application.properties
@@ -11,6 +11,7 @@ spring.config.import=classpath:i18n/messages_${fr.insee.kraftwerk.lang}.properti
 fr.insee.kraftwerk.version=@project.version@
 springdoc.swagger-ui.path=/index.html
 springdoc.api-docs.resolve-schema-properties=true
+springdoc.swagger-ui.tagsSorter=alpha
 
 
 fr.insee.postcollecte.csv.output.quote ="
