+++ kraftwerk-api/src/main/resources/i18n/messages_en.properties
@@ -2,6 +2,9 @@
 ## ENGLISH DICTIONARY	##
 ##########################################
 
+tag.main= 1- Main services
+tag.stepbystep= 2- Step by step services
+
 summary.main=Main service : call all steps
 description.main=This service does : \n  - buildVtlBindings \n - unimodalProcessing \n - multimodalProcessing  \n - Separate the different information's levels and export
