+++ kraftwerk-api/src/main/resources/i18n/messages_fr.properties
@@ -2,6 +2,10 @@
 ## FRENCH DICTIONARY	##
 ##########################################
 
+tag.main= 1- Services principaux
+tag.stepbystep= 2- Services par étapes
+
+
 summary.main=Service principal : ce service appelle toutes les étapes successivement
 description.main=Ce service enchaîne les traitements : \n  - buildVtlBindings : construction des jeux de données au format VTL \n - unimodalProcessing : traitement unimodal pour chaque mode de collecte disponible \n - multimodalProcessing : Réconciliation des différents modes et traitements multimodal \n - Séparation en tables par niveau d'information (boucles) et export CSV
